ai-backend/pyproject.toml
[tool.poetry]
name = "ai-backend"
version = "0.1.0"
description = ""
authors = ["garv-2501 <garvsinghrawlot1.0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = {extras = ["asyncio"], version = "^5.11.4"}
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
langchain = "^0.3.1"
langchain-openai = "^0.2.1"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pymongo = {extras = ["srv"], version = "^4.10.0"}
langchain-community = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
